unreel-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
google-generativeai>=0.8.5
yt-dlp==2023.11.16
ffmpeg-python==0.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
openai-whisper==20250625